Create 03_Fraction Class
Implement Fraction Class with the following properties and functions.
Properties:
1. Numerator
2.Denominator
Functions: 1. Parametrized constructor-
Sets the numerator and denominator values.
2. add-
This function adds the two fraction following the adding up of two fraction rules and updates the first fractional number.
e.g.
if f1 = 1/4 and f2 = 3/5
f1.add(f2); results in:
f1 = 17/20 and f2 = 3/5
3.multiply-
This function multiplies the two fraction and updates the first fractional number
e.g.
if f1 = 1/4 and f2 = 3/5
f1.multiply(f2); results in:
f1 = 3/20 and f2 = 3/5
4.simplify
This function simplifies the fractional value. Use inbuilt_gcd() function.
e.g.
if f1=5/20. which can be further simplified as 1/4 so this if f1-5/20, which can be further simplified as 1/4 so this simplify will perform this.
8. print() -
Prints the final answer in numerator/denominator form.

Input Format:
The first line of input contains two space-separated integer values, which represent numerator and denominator.
The second line of input contains an integer N representing queries to define how many times the operation must occur.
Next, N lines contain 3 space-separated integers representing the type of query.
numerator, and denominator.

Output Format:
Print the value of the fraction after every query in a separate line.

Input Constraints:
1 <= numerator <= 100
1 <-denominator <- 100
1 <= N <= 15
type - 1 or 2