Create 06_Exception_Calculate Power
Given a class 'Solution', you must implement a single method, 'power(string, string)', which takes two strings, 'A' and 'B'. You must return 'A^B'
Throw and catch the following exceptions with a handler message:
In case of an Arithmetic exception, return 'Arithmetic Exception occurred.. If 'A' or 'B' is negative, return 'A or B can't be negative..
If 'A' and 'B' are zero, return 'A and B both can't be zero'.
If Number Format Exception occurred, return "Number Format Exception occurred.'.
If another error occurs, return 'Some other error occurred..
Note:
For the second and third exceptions, throw user-defined exceptions.
Example:
Input:
A=-1, B=4
Output:
A or B can't be negative.
Explanation: Since 'A' is negative, we return 'A or B can't be negative..
Detailed explanation (Input/output format, Notes, Images) Input Format:
The first line of the input contains two strings, 'A' and 'B'.
Output Format:-
Output is printed on a separate line.

Note:-
You don't need to print anything. Just implement the given function.
Constraints:
The answer will fit in a 32-bit integer.
Time Limit: 1 sec