Create 08_Exception_Maximum Of All
Given a class 'Solution', you must implement a single method, maximumOfAll(string, string)', which takes two strings, 'A' and 'B'. You must calculate 'A+B', 'A-B', 'A/B', and 'A'B' and return the maximum of all the values.
In case of an exception:
If an Arithmetic exception occurred, return 'Arithmetic Exception occurred.. If Number Format Exception occurred, return "Number Format Exception occurred..
If another error occurs, return "Some other error occurred..
Example:
Input:
A=3. B=2
Output:
6
Explanation: A+B = 5, A-B = 1, A/B = 1, A*B = 6.
Hence we return 6.

Input Format:
The first line of the input contains two strings, 'A' and 'B'.

Output Format:-
Output is printed on a separate line.