Create 09_Collection_Sharing_Coins
You have infinite coins and your friends have none. You are willing to share your coins. You have to process 'Q' queries of 2 types as following:
Each query contains an integer 't', denoting the type of the query.
Type 1: This query contains a string 'p' and an integer 'c', the name of the friend you are giving coins to, and the amount of coins you are giving to this friend. Type 2: This query contains a string 'p' which is a name. Print the number of coins this person has.
For Example:
For queries - [ [1, "ninja", 10], [1, "sam", 20], [1, "ninja", 15], [2, "ninja"], [2, "jack"], [1, "ninja", 15] ].
Answer = [25,0]
Since we have only given '25' coins to "ninja" until the query [2, "ninja"] is asked, the answer to it is '25', and we haven't given any coins to "jack", so the answer to query [2. "jack"] it is '0'.

Input format:
The first line contains an integer 'Q', representing the number of queries.
The next 'Q' lines contain the description of queries:
For each query:
The first line contains an integer 't', representing the type of query.
For Type 1 queries: The first line also contains a string 'p' and an integer 'c', name of the 'friend' and number of 'coins' given to the 'friend".
For Type 2 queries:
The first line also contains a string 'p', the name of the 'friend'.

Output Format:
Return an array containing answers to query of type 2 in order.

Constraints:
1<= 'Q' <= 10
1<= 'p.length <= 10
1 <= 'c' <= 100
Time Limit: 1 sec